delete image when product is deleted

+++ 7. Authentication/controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require('../util/file');
 const {validationResult} = require('express-validator');
 
 const Product = require('../models/product');
@@ -156,6 +157,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if(updatedImage){
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = updatedImage.path;
             }
             // product.imageUrl = updatedImageUrl;
@@ -194,14 +196,21 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-        const error = new Error(err);
-        error.httpStatusCode = 500;
-        return next(error);
-    });
+  Product.findById(prodId)
+      .then(product => {
+          if(!product){
+              return next(new Error('Product not found.'));
+          }
+          fileHelper.deleteFile(product.imageUrl);
+          return Product.deleteOne({_id: prodId, userId: req.user._id})
+      })
+      .then(() => {
+          console.log('DESTROYED PRODUCT');
+          res.redirect('/admin/products');
+      })
+      .catch(err => {
+          const error = new Error(err);
+          error.httpStatusCode = 500;
+          return next(error);
+      });
 };
